+++ apps/desktop/src/sshEnvironment.ts
@@ -13,18 +13,34 @@ import type {
   DesktopSshEnvironmentBootstrap,
   DesktopSshEnvironmentTarget,
   DesktopSshPasswordPromptRequest,
-  DesktopUpdateChannel,
   ExecutionEnvironmentDescriptor,
 } from "@t3tools/contracts";
+import {
+  DEFAULT_REMOTE_PORT,
+  baseSshArgs,
+  buildSshHostSpec,
+  collectSshConfigAliasesFromFile,
+  discoverSshHosts,
+  getLastNonEmptyOutputLine,
+  isSshAuthFailure,
+  normalizeSshErrorMessage,
+  parseKnownHostsHostnames,
+  parseSshResolveOutput,
+  remoteStateKey,
+  resolveRemoteT3CliPackageSpec,
+  resolveSshConfigIncludePattern,
+  targetConnectionKey,
+} from "@t3tools/ssh";
+import { Effect } from "effect";
 
 import { waitForHttpReady } from "./backendReadiness.ts";
 
-const DEFAULT_REMOTE_PORT = 3773;
+export { resolveRemoteT3CliPackageSpec };
+
 const REMOTE_PORT_SCAN_WINDOW = 200;
 const SSH_ASKPASS_DIR_NAME = "t3code-ssh-askpass";
 const TUNNEL_SHUTDOWN_TIMEOUT_MS = 2_000;
 const SSH_READY_TIMEOUT_MS = 20_000;
-const PUBLISHABLE_T3_VERSION_PATTERN = /^\d+\.\d+\.\d+(?:-[0-9A-Za-z.-]+)?$/u;
 
 const DISCOVER_SSH_HOSTS_CHANNEL = "desktop:discover-ssh-hosts";
 const ENSURE_SSH_ENVIRONMENT_CHANNEL = "desktop:ensure-ssh-environment";
@@ -80,199 +96,6 @@ interface DesktopSshEnvironmentManagerOptions {
   readonly resolveCliPackageSpec?: () => string;
 }
 
-const NO_HOSTS = [] as const;
-
-function stripInlineComment(line: string): string {
-  const hashIndex = line.indexOf("#");
-  return (hashIndex >= 0 ? line.slice(0, hashIndex) : line).trim();
-}
-
-function splitDirectiveArgs(value: string): ReadonlyArray<string> {
-  return value
-    .trim()
-    .split(/\s+/u)
-    .map((entry) => entry.trim())
-    .filter((entry) => entry.length > 0);
-}
-
-function expandHomePath(input: string, homeDir: string = OS.homedir()): string {
-  return input.replace(/^~(?=$|\/|\\)/u, homeDir);
-}
-
-function resolveSshConfigIncludePattern(
-  includePattern: string,
-  _directory: string,
-  homeDir: string = OS.homedir(),
-): string {
-  const expandedPattern = expandHomePath(includePattern, homeDir);
-  return Path.isAbsolute(expandedPattern)
-    ? expandedPattern
-    : Path.resolve(Path.join(homeDir, ".ssh"), expandedPattern);
-}
-
-function hasSshPattern(value: string): boolean {
-  return value.includes("*") || value.includes("?") || value.startsWith("!");
-}
-
-function escapeRegex(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/gu, "\\$&");
-}
-
-function globToRegExp(pattern: string): RegExp {
-  return new RegExp(
-    `^${escapeRegex(pattern).replace(/\\\*/gu, ".*").replace(/\\\?/gu, ".")}$`,
-    "u",
-  );
-}
-
-function expandGlob(pattern: string): ReadonlyArray<string> {
-  if (!pattern.includes("*") && !pattern.includes("?")) {
-    return FS.existsSync(pattern) ? [pattern] : NO_HOSTS;
-  }
-
-  const directory = Path.dirname(pattern);
-  const basePattern = Path.basename(pattern);
-  if (!FS.existsSync(directory)) {
-    return NO_HOSTS;
-  }
-
-  const matcher = globToRegExp(basePattern);
-  return FS.readdirSync(directory)
-    .filter((entry) => matcher.test(entry))
-    .map((entry) => Path.join(directory, entry))
-    .filter((entry) => FS.existsSync(entry))
-    .toSorted((left, right) => left.localeCompare(right));
-}
-
-function collectSshConfigAliasesFromFile(
-  filePath: string,
-  visited = new Set<string>(),
-  homeDir: string = OS.homedir(),
-): ReadonlyArray<string> {
-  const resolvedPath = Path.resolve(filePath);
-  if (visited.has(resolvedPath) || !FS.existsSync(resolvedPath)) {
-    return NO_HOSTS;
-  }
-  visited.add(resolvedPath);
-
-  const aliases = new Set<string>();
-  const directory = Path.dirname(resolvedPath);
-  const raw = FS.readFileSync(resolvedPath, "utf8");
-
-  for (const line of raw.split(/\r?\n/u)) {
-    const stripped = stripInlineComment(line);
-    if (stripped.length === 0) {
-      continue;
-    }
-
-    const [directive = "", ...rawArgs] = splitDirectiveArgs(stripped);
-    const normalizedDirective = directive.toLowerCase();
-    if (normalizedDirective === "include") {
-      for (const includePattern of rawArgs) {
-        const resolvedPattern = resolveSshConfigIncludePattern(includePattern, directory, homeDir);
-        for (const includedPath of expandGlob(resolvedPattern)) {
-          for (const alias of collectSshConfigAliasesFromFile(includedPath, visited, homeDir)) {
-            aliases.add(alias);
-          }
-        }
-      }
-      continue;
-    }
-
-    if (normalizedDirective !== "host") {
-      continue;
-    }
-
-    for (const alias of rawArgs) {
-      if (alias.length === 0 || hasSshPattern(alias)) {
-        continue;
-      }
-      aliases.add(alias);
-    }
-  }
-
-  return [...aliases].toSorted((left, right) => left.localeCompare(right));
-}
-
-function normalizeKnownHostsHostname(rawHost: string): string {
-  const bracketMatch = /^\[([^\]]+)\]:(\d+)$/u.exec(rawHost);
-  if (bracketMatch?.[1]) {
-    return bracketMatch[1];
-  }
-
-  if (!rawHost.includes(":")) {
-    return rawHost;
-  }
-
-  const firstColonIndex = rawHost.indexOf(":");
-  const lastColonIndex = rawHost.lastIndexOf(":");
-  return firstColonIndex === lastColonIndex ? rawHost.slice(0, lastColonIndex) : rawHost;
-}
-
-function parseKnownHostsHostnames(raw: string): ReadonlyArray<string> {
-  const hostnames = new Set<string>();
-
-  for (const line of raw.split(/\r?\n/u)) {
-    const trimmed = line.trim();
-    if (trimmed.length === 0 || trimmed.startsWith("#")) {
-      continue;
-    }
-
-    const withoutMarker = trimmed.startsWith("@")
-      ? trimmed.split(/\s+/u).slice(1).join(" ")
-      : trimmed;
-    const [hostField = ""] = withoutMarker.split(/\s+/u);
-    if (hostField.length === 0 || hostField.startsWith("|")) {
-      continue;
-    }
-
-    for (const rawHost of hostField.split(",")) {
-      const host = normalizeKnownHostsHostname(rawHost).trim();
-      if (host.length === 0 || hasSshPattern(host)) {
-        continue;
-      }
-      hostnames.add(host);
-    }
-  }
-
-  return [...hostnames].toSorted((left, right) => left.localeCompare(right));
-}
-
-function readKnownHostsHostnames(filePath: string): ReadonlyArray<string> {
-  if (!FS.existsSync(filePath)) {
-    return NO_HOSTS;
-  }
-
-  return parseKnownHostsHostnames(FS.readFileSync(filePath, "utf8"));
-}
-
-function parseSshResolveOutput(alias: string, stdout: string): DesktopSshEnvironmentTarget {
-  const values = new Map<string, string>();
-  for (const line of stdout.split(/\r?\n/u)) {
-    const trimmed = line.trim();
-    if (trimmed.length === 0) {
-      continue;
-    }
-    const [key, ...rest] = trimmed.split(/\s+/u);
-    if (!key || rest.length === 0 || values.has(key)) {
-      continue;
-    }
-    values.set(key, rest.join(" ").trim());
-  }
-
-  const hostname = values.get("hostname")?.trim() || alias;
-  const username = values.get("user")?.trim() || null;
-  const rawPort = values.get("port")?.trim() ?? "";
-  const parsedPort = Number.parseInt(rawPort, 10);
-
-  return {
-    alias,
-    hostname,
-    username,
-    port: Number.isInteger(parsedPort) ? parsedPort : null,
-  };
-}
-
 async function findAvailableLocalPort(): Promise<number> {
   return await new Promise<number>((resolve, reject) => {
     const server = Net.createServer();
@@ -294,22 +117,6 @@ async function findAvailableLocalPort(): Promise<number> {
       });
     });
   });
-}
-
-function targetConnectionKey(target: DesktopSshEnvironmentTarget): string {
-  return `${target.alias}\u0000${target.hostname}\u0000${target.username ?? ""}\u0000${target.port ?? ""}`;
-}
-
-function remoteStateKey(target: DesktopSshEnvironmentTarget): string {
-  return Crypto.createHash("sha256").update(targetConnectionKey(target)).digest("hex").slice(0, 16);
-}
-
-function buildSshHostSpec(target: DesktopSshEnvironmentTarget): string {
-  const destination = target.alias.trim() || target.hostname.trim();
-  if (destination.length === 0) {
-    throw new Error("SSH target is missing its alias/hostname.");
-  }
-  return target.username ? `${target.username}@${destination}` : destination;
 }
 
 function getDefaultSshAskpassDirectory(): string {
@@ -450,36 +257,6 @@ function buildSshChildEnvironment(input?: {
   };
 }
 
-function baseSshArgs(
-  target: DesktopSshEnvironmentTarget,
-  input?: { readonly batchMode?: "yes" | "no" },
-): string[] {
-  return [
-    "-o",
-    `BatchMode=${input?.batchMode ?? "no"}`,
-    "-o",
-    "ConnectTimeout=10",
-    ...(target.port !== null ? ["-p", String(target.port)] : []),
-  ];
-}
-
-function normalizeSshErrorMessage(stderr: string, fallbackMessage: string): string {
-  const cleaned = stderr.trim();
-  return cleaned.length > 0 ? cleaned : fallbackMessage;
-}
-
-function isSshAuthFailure(error: unknown): boolean {
-  const message = error instanceof Error ? error.message : String(error);
-  const normalized = message.toLowerCase();
-  return (
-    /permission denied \((?:publickey|password|keyboard-interactive|hostbased|gssapi-with-mic)[^)]*\)/u.test(
-      normalized,
-    ) ||
-    /authentication failed/u.test(normalized) ||
-    /too many authentication failures/u.test(normalized)
-  );
-}
-
 async function runSshCommand(
   target: DesktopSshEnvironmentTarget,
   input?: {
@@ -594,33 +371,6 @@ function buildRemoteLaunchScript(input?: { readonly packageSpec?: string }): str
     T3_DEFAULT_REMOTE_PORT: String(DEFAULT_REMOTE_PORT),
     T3_REMOTE_PORT_SCAN_WINDOW: String(REMOTE_PORT_SCAN_WINDOW),
   });
-}
-
-function getLastNonEmptyOutputLine(stdout: string): string | null {
-  return (
-    stdout
-      .trim()
-      .split(/\r?\n/u)
-      .map((entry) => entry.trim())
-      .findLast((entry) => entry.length > 0) ?? null
-  );
-}
-
-export function resolveRemoteT3CliPackageSpec(input: {
-  readonly appVersion: string;
-  readonly updateChannel: DesktopUpdateChannel;
-  readonly isDevelopment?: boolean;
-}): string {
-  const appVersion = input.appVersion.trim();
-  if (!input.isDevelopment && PUBLISHABLE_T3_VERSION_PATTERN.test(appVersion)) {
-    return `t3@${appVersion}`;
-  }
-
-  if (input.isDevelopment) {
-    return "t3@nightly";
-  }
-
-  return input.updateChannel === "nightly" ? "t3@nightly" : "t3@latest";
 }
 
 function buildRemoteT3RunnerScript(input?: { readonly packageSpec?: string }): string {
@@ -769,40 +519,7 @@ async function stopTunnel(entry: SshTunnelEntry): Promise<void> {
 export async function discoverDesktopSshHosts(input?: {
   readonly homeDir?: string;
 }): Promise<readonly DesktopDiscoveredSshHost[]> {
-  const homeDir = input?.homeDir ?? OS.homedir();
-  const sshDirectory = Path.join(homeDir, ".ssh");
-  const configAliases = collectSshConfigAliasesFromFile(
-    Path.join(sshDirectory, "config"),
-    new Set<string>(),
-    homeDir,
-  );
-  const knownHosts = readKnownHostsHostnames(Path.join(sshDirectory, "known_hosts"));
-  const discovered = new Map<string, DesktopDiscoveredSshHost>();
-
-  for (const alias of configAliases) {
-    discovered.set(alias, {
-      alias,
-      hostname: alias,
-      username: null,
-      port: null,
-      source: "ssh-config",
-    });
-  }
-
-  for (const hostname of knownHosts) {
-    if (discovered.has(hostname)) {
-      continue;
-    }
-    discovered.set(hostname, {
-      alias: hostname,
-      hostname,
-      username: null,
-      port: null,
-      source: "known-hosts",
-    });
-  }
-
-  return [...discovered.values()].toSorted((left, right) => left.alias.localeCompare(right.alias));
+  return await Effect.runPromise(discoverSshHosts(input));
 }
 
 export class DesktopSshEnvironmentManager {
@@ -1369,8 +1086,6 @@ export const __test = {
   getLastNonEmptyOutputLine,
   isSshAuthFailure,
   collectSshConfigAliasesFromFile,
-  expandHomePath,
-  normalizeKnownHostsHostname,
   parseKnownHostsHostnames,
   parseSshResolveOutput,
   readSshScriptTemplate,
